Use size_t for size parameters in cilk merge
In some cases int could overflow, causing a segmentation fault.

+++ 2-merge/merge.c
@@ -2,14 +2,14 @@
 
 #include <cilk.h>
 
-extern void EXPORT(merge_impl) (CilkContext *context, const TYPE *a, int n,
-        const TYPE *b, int m, TYPE *c, perf_t *perf, int *nproc);
+extern void EXPORT(merge_impl) (CilkContext *context, const TYPE *a, size_t n,
+        const TYPE *b, size_t m, TYPE *c, perf_t *perf, int *nproc);
 
-void merge_seq(const TYPE *a, int n, const TYPE *b, int m, TYPE *c)
+void merge_seq(const TYPE *a, size_t n, const TYPE *b, size_t m, TYPE *c)
 {
-    int i = 0;
-    int j = 0;
-    int k = 0;
+    size_t i = 0;
+    size_t j = 0;
+    size_t k = 0;
 
     while (i < n && j < m) {
         c[k++] = ((a[i] < b[j]) ? a[i++] : b[j++]);
@@ -19,7 +19,7 @@ void merge_seq(const TYPE *a, int n, const TYPE *b, int m, TYPE *c)
     while (j < m) c[k++] = b[j++];
 }
 
-TYPE *merge_ref(const TYPE *a, int n, const TYPE *b, int m)
+TYPE *merge_ref(const TYPE *a, size_t n, const TYPE *b, size_t m)
 {
     TYPE *c = calloc(n + m, sizeof(TYPE));
     if (c == NULL) {
@@ -31,7 +31,7 @@ TYPE *merge_ref(const TYPE *a, int n, const TYPE *b, int m)
     return c;
 }
 
-TYPE *merge(const TYPE *a, int n, const TYPE *b, int m, perf_t *perf, const char *nproc,
+TYPE *merge(const TYPE *a, size_t n, const TYPE *b, size_t m, perf_t *perf, const char *nproc,
         int *nproc_ret) {
     const char *argv[] = {algorithm_name, "--nproc", nproc, 0};
     int argc = 3;
